[refactor/21] 자녀 생성후, 진단완료시 자녀의 임베딩값 저장
